Dev: qdevice: Consolidate qdevice related separate functions into qdevice.QDevice class
